task 02 - criação do hook useUser e melhoramento do código
