+++ src/components/InputText/InputText.tsx
@@ -2,12 +2,13 @@
 import { TextField } from "@mui/material";
 import { Controller, FieldValues, Path, useFormContext } from "react-hook-form";
 
-type Props<T extends FieldValues> = {
+type InputTextProps<T extends FieldValues> = {
+  id: string,
   name: Path<T>,
   label: string,
 }
 
-export default function InputText<T extends FieldValues>({ name, label }: Props<T>) {
+export default function InputText<T extends FieldValues>({ id, name, label }: InputTextProps<T>) {
   const { control } = useFormContext()
   return (
     <Controller
@@ -15,6 +16,7 @@ export default function InputText<T extends FieldValues>({ name, label }: Props<
       name={name}
       render={({ field, fieldState: { error } }) => {
         return <TextField
+          id={id}
           label={label}
           {...field}
           error={!!error}
