+++ src/types/types.ts
@@ -1,0 +1,4 @@
+import { userSchema } from "@/model/model";
+import { z } from "zod";
+
+export type UserSchema = z.infer<typeof userSchema>
